Rename 'mock-data' files so they don't have the same file name
